requirements.txt
pandas>=1.3.0,<2.0.0
PyPDF2>=3.0.0,<4.0.0
tqdm>=4.65.0,<5.0.0
python-magic>=0.4.27,<0.5.0
beautifulsoup4>=4.12.0,<5.0.0
lxml>=4.9.0,<5.0.0 